* commit 'e1e814c7c640a51511b149985d57f8bfd4a8bf78':
  MTP: add strict bounds checking for all incoming packets

+++ media/mtp/MtpDevice.h
@@ -98,7 +98,7 @@ public:
     bool                    readObject(MtpObjectHandle handle,
                                     bool (* callback)(void* data, int offset,
                                             int length, void* clientData),
-                                    int objectSize, void* clientData);
+                                    size_t objectSize, void* clientData);
     bool                    readObject(MtpObjectHandle handle, const char* destPath, int group,
                                     int perm);
